Workaround some icc issues.
The intel compiler has an ICE when explicitly initializing fields that are
arrays of std::atomic. This occurs in the process Bitmap, so we just leave that
field implicitly initialized.

The intel compiler has issues with the stdatomic.h header that I was including
in some unit tests post-libsync removal, so I switched those tests to compile
with C++ and to use C++ atomics.

+++ tests/unit/thread_yield.cpp
@@ -60,7 +60,7 @@ static HPX_ACTION(HPX_DEFAULT, HPX_MARSHALLED, _yield_worker,
                   _yield_worker_handler, HPX_POINTER, HPX_SIZE_T);
 
 static int thread_yield_handler(void) {
-  int num_threads = hpx_get_num_threads();
+  unsigned num_threads = (unsigned)hpx_get_num_threads();
   std::atomic<size_t> counter;
 
   struct _yield_args args = {
